added functions for cancel check

+++ database/migrations/2025_11_17_073355_create_borrowed_checks_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('borrowed_checks', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnUpdate()->cascadeOnDelete();
-            $table->integer('check_id');
+            $table->unsignedBigInteger('check_id');
             $table->string('name');
             $table->string('reason');
             $table->enum('check', ['cv', 'crf']);
